Use "client API" throughtout the docs
Historically, we have used "client API" but with the dev-site update,
"API clients" started showing up in several places. This is
inconsistent with our other usage and it is particularly confusing
for the database APIs. The naming probably could be better generally,
e.g., "database libraries" for the MMDB readers and "client libaries"
for the web service client libraries, but this at least brings
consistency with our usage elsewhere.